% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IVM.R
\name{IVM.regression}
\alias{IVM.regression}
\title{Perform IVM selection for sparse GP regression}
\usage{
IVM.regression(predictors, responses, activeSetSize, variance, kernel.function)
}
\arguments{
\item{predictors}{a numeric matrix of predictors (X)}

\item{responses}{a numeric vector of responses (Y)}

\item{activeSetSize}{the number of points to select for the active set}

\item{variance}{the variance of the Gaussian noise function}

\item{kernel.function}{a positive definite function which takes two vectors
x_1 and x_2 and returns the covariance between the two points}
}
\value{
An object of class IVM. Active set indices are stored in activeSet
}
\description{
Perform IVM selection for sparse GP regression
}

